refactor mememaker into component
